# opt:(encoder) use std `strconv.AppendInt` for better performance on arm