#### What type of PR is this?
<!--
Add one of the following kinds:

build: Changes that affect the build system or external dependencies (example scopes: gulp, broccoli, npm)
ci: Changes to our CI configuration files and scripts (example scopes: Travis, Circle, BrowserStack, SauceLabs)
docs: Documentation only changes
feat: A new feature
optimize: A new optimization
fix: A bug fix
perf: A code change that improves performance
refactor: A code change that neither fixes a bug nor adds a feature
style: Changes that do not affect the meaning of the code (white space, formatting, missing semi-colons, etc)
test: Adding missing tests or correcting existing tests
chore: Changes to the build process or auxiliary tools and libraries such as documentation generation
-->

#### Check the PR title.
<!--
The description of the title will be attached in Release Notes, 
so please describe it from user-oriented, what this PR does / why we need it.
Please check your PR title with the below requirements:
-->
- [ ] This PR title match the format: \<type\>(optional scope): \<description\>
- [ ] The description of this PR title is user-oriented and clear enough for others to understand.
- [ ] Attach the PR updating the user documentation if the current PR requires user awareness at the usage level. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)


#### (Optional) Translate the PR title into Chinese.


#### (Optional) More detailed description for this PR(en: English/zh: Chinese).
<!--
Provide more detailed info for review(e.g., it's recommended to provide perf data if this is a perf type PR).
-->
en: according to benchmark, `strconv.AppendItn()` has better performance than native implementation in `alg.I64toa`:
```
goos: darwin
goarch: arm64
pkg: github.com/bytedance/sonic/internal/encoder/alg
cpu: Apple M3 Pro
BenchmarkI64toa/sonic-1-12              48961936                22.01 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-1-12                329461662                3.600 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-16-12             55322376                21.48 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-16-12               438972664                2.730 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-256-12            55268763                21.42 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-256-12              205196311                5.884 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-4096-12           54943064                21.90 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-4096-12             207751256                5.740 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-65536-12          54216253                21.70 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-65536-12            175604583                6.849 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-1048576-12        53292178                22.54 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-1048576-12          134011879                8.747 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-16777216-12       53392560                22.75 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-16777216-12         136341493                9.195 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-268435456-12      47283418                24.78 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-268435456-12        125863491                9.415 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-4294967296-12             46340396                24.68 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-4294967296-12               125847438                9.580 ns/op           0 B/op          0 allocs/op
BenchmarkI64toa/sonic-68719476736-12            48397423                24.95 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-68719476736-12              100000000               10.32 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/sonic-1099511627776-12          48338691                24.72 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-1099511627776-12            100000000               11.15 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/sonic-17592186044416-12         47782588                25.10 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-17592186044416-12           100000000               11.25 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/sonic-281474976710656-12        50566502                24.11 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-281474976710656-12          100000000               11.70 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/sonic-4503599627370496-12       45584118                24.80 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-4503599627370496-12         96807373                12.33 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/sonic-72057594037927936-12      49537647                23.57 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-72057594037927936-12        95043858                12.13 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/sonic-1152921504606846976-12            49055928                25.24 ns/op            0 B/op          0 allocs/op
BenchmarkI64toa/std-1152921504606846976-12              91301910                12.96 ns/op            0 B/op          0 allocs/op
```
zh(optional): 

#### (Optional) Which issue(s) this PR fixes:
<!--
Automatically closes linked issue when PR is merged.
Eg: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

#### (optional) The PR that updates user documentation:
<!--
If the current PR requires user awareness at the usage level, please submit a PR to update user docs. [User docs repo](https://github.com/cloudwego/cloudwego.github.io)
-->